% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels-neonormal.R
\name{NeoNormalKernels}
\alias{NeoNormalKernels}
\alias{dMSNBurr_k}
\alias{rMSNBurr_k}
\alias{dMSNBurr2a_k}
\alias{rMSNBurr2a_k}
\alias{dGMSNBurr_k}
\alias{rGMSNBurr_k}
\alias{dSEP_k}
\alias{rSEP_k}
\alias{dLEP_k}
\alias{rLEP_k}
\alias{dFSSN_k}
\alias{rFSSN_k}
\alias{dFOSSEP_k}
\alias{rFOSSEP_k}
\alias{dFSST_k}
\alias{rFSST_k}
\alias{dJFST_k}
\alias{rJFST_k}
\alias{dSkewMvN_k}
\alias{rSkewMvN_k}
\alias{dSkewMvNO_k}
\alias{rSkewMvNO_k}
\alias{dSkewMvNOG_k}
\alias{dSkewMvIT_k}
\alias{rSkewMvIT_k}
\alias{dSkewMvITO_k}
\alias{rSkewMvITO_k}
\alias{dSkewMvITOG_k}
\title{Scalar neo-normal densities for NIMBLE}
\usage{
dMSNBurr_k(x, mu, sigma, alpha, log = 0)

rMSNBurr_k(n, mu, sigma, alpha)

dMSNBurr2a_k(x, mu, sigma, alpha, log = 0)

rMSNBurr2a_k(n, mu, sigma, alpha)

dGMSNBurr_k(x, mu, sigma, alpha, theta, log = 0)

rGMSNBurr_k(n, mu, sigma, alpha, theta)

dSEP_k(x, mu, sigma, nu, log = 0)

rSEP_k(n, mu, sigma, nu)

dLEP_k(x, mu, sigma, nu, log = 0)

rLEP_k(n, mu, sigma, nu)

dFSSN_k(x, mu, sigma, alpha, log = 0)

rFSSN_k(n, mu, sigma, alpha)

dFOSSEP_k(x, mu, sigma, alpha, theta, log = 0)

rFOSSEP_k(n, mu, sigma, alpha, theta)

dFSST_k(x, mu, sigma, alpha, nu, log = 0)

rFSST_k(n, mu, sigma, alpha, nu)

dJFST_k(x, mu, sigma, alpha, theta, log = 0)

rJFST_k(n, mu, sigma, alpha, theta)

dSkewMvN_k(x, mu, Sigma, gam, log = 0)

rSkewMvN_k(n, mu, Sigma, gam)

dSkewMvNO_k(x, mu, Sigma, gam, theta, log = 0)

rSkewMvNO_k(n, mu, Sigma, gam, theta)

dSkewMvNOG_k(x, mu, Sigma, gam, theta, log = 0)

dSkewMvIT_k(x, mu, Sigma, gam, nu, log = 0)

rSkewMvIT_k(n, mu, Sigma, gam, nu)

dSkewMvITO_k(x, mu, Sigma, gam, nu, theta, log = 0)

rSkewMvITO_k(n, mu, Sigma, gam, nu, theta)

dSkewMvITOG_k(x, mu, Sigma, gam, nu, theta, log = 0)
}
\arguments{
\item{x}{Observation.}

\item{mu}{Location.}

\item{sigma}{Scale.}

\item{alpha}{Shape.}

\item{log}{Return the log density.}

\item{n}{Number of draws (required by NIMBLE; always 1).}
}
\value{
The (log) density, or a draw.
}
\description{
Custom NIMBLE distributions for the neo-normal families, used both as
mixture component densities and as the emission primitives called by the
hidden Markov forward kernels.
}
\details{
\code{dMSNBurr_k} implements the MSNBurr density (Iriawan, 2000; Choir,
2020), parameterised by location \code{mu}, scale \code{sigma} and shape
\code{alpha}, where \code{alpha = 1} recovers symmetry and the normalising
constant \code{omega} is computed on the log scale for stability at extreme
\code{alpha}.

These are exported because NIMBLE resolves user-defined distributions by
name on the search path when generating C++ code. They are not intended to
be called directly.
}
\references{
Iriawan, N. (2000). \emph{Computationally intensive approaches
  to inference in neo-normal linear models}. PhD thesis, Curtin University.

  Choir, A.S. (2020). \emph{The new neo-normal distributions and their
  properties}. Dissertation, Institut Teknologi Sepuluh Nopember.
  ISBN:978-602-5539-45-4
}
\keyword{internal}
