% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnimix-package.R
\docType{package}
\name{Rnimix-package}
\alias{Rnimix}
\alias{Rnimix-package}
\title{Rnimix: Bayesian Mixture Clustering and Regression with NIMBLE}
\description{
Bayesian mixture modelling built on the \pkg{nimble} platform. The package
implements univariate and multivariate mixture clustering and
mixture-of-regressions through two inference engines: a Dirichlet Process
Mixture (DPM) engine based on the Chinese Restaurant Process (which estimates
the number of occupied components) and a fixed-K finite-mixture engine. It is
organised around an extensible S4 \code{\linkS4class{DistributionSpec}}
contract so that new component distributions and engines can be added without
rewriting existing code.
}
\section{Inference engines}{

\itemize{
  \item \code{method = "dpm"}: Dirichlet process / Chinese restaurant process;
    the number of occupied components is estimated from the data.
  \item \code{method = "fixedk"}: finite mixture with a known number of
    components \code{K}.
}
}

\section{Component distributions}{

Gaussian (univariate and multivariate), Student-t and Normal-Gamma
(heavy-tailed, univariate and multivariate), and Poisson / Binomial counts,
for both clustering (\code{\link{nimixClust}}) and regression
(\code{\link{nimixReg}}, including multivariate responses).
}

\references{
de Valpine, P., Turek, D., Paciorek, C.J., Anderson-Bergman, C.,
Temple Lang, D., & Bodik, R. (2017). Programming with models: writing
statistical algorithms for general model structures with NIMBLE.
\emph{Journal of Computational and Graphical Statistics}, 26(2), 403--413.
\doi{10.1080/10618600.2016.1172487}

Neal, R.M. (2000). Markov chain sampling methods for Dirichlet process
mixture models. \emph{Journal of Computational and Graphical Statistics},
9(2), 249--265. \doi{10.1080/10618600.2000.10474879}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/madsyair/Rnimix}
  \item \url{https://madsyair.github.io/Rnimix/}
  \item Report bugs at \url{https://github.com/madsyair/Rnimix/issues}
}

}
\author{
\strong{Maintainer}: Achmad Syahrul Choir \email{madsyair@stis.ac.id} (\href{https://orcid.org/0000-0001-7088-0646}{ORCID})

Authors:
\itemize{
  \item Achmad Syahrul Choir \email{madsyair@stis.ac.id} (\href{https://orcid.org/0000-0001-7088-0646}{ORCID})
}

}
\keyword{internal}
